支持int64

+++ v8vm/gosrc/export_to_c.go
@@ -3,8 +3,8 @@ package v8vm
 /*
 #include <stdio.h>
 
-extern int BalanceTransfer(__int64, char*, char*, int);
-int BalanceTransferStub(__int64 vmid, char* from, char* to, int amount)
+extern int BalanceTransfer(__int64, char*, char*, __int64);
+int BalanceTransferStub(__int64 vmid, char* from, char* to, __int64 amount)
 {
 	return BalanceTransfer(vmid, from, to, amount);
 }
