+++ v8vm/gosrc/stub_for_c.go
@@ -8,7 +8,7 @@ import "C"
 import "fmt"
 
 //export BalanceTransfer
-func BalanceTransfer(vmid int64, from *C.char, to *C.char, amount int32) _Ctype_int {
+func BalanceTransfer(vmid int64, from *C.char, to *C.char, amount int64) _Ctype_int {
 	fmt.Println("!!!!!!!!!! BalanceTransfer(), vmid =", vmid, ", from =", C.GoString(from), ", to =", C.GoString(to), ", value2 =", amount)
 	return 0
 }
